sistema_os_melhorado_pacote/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-Limiter==3.3.1
Flask-Caching==2.0.2
Flask-Compress==1.13
Flask-Cors==4.0.0
SQLAlchemy==2.0.9
Werkzeug==2.2.3
WTForms==3.0.1
email-validator==2.0.0
python-dotenv==1.0.0
Pillow==9.5.0
PyMySQL==1.0.3
cryptography==40.0.2
bcrypt==4.0.1
pdfkit==1.0.0
WeasyPrint==59.0
openpyxl==3.1.2
matplotlib==3.7.1
pandas==2.0.0
pytest==7.3.1
pytest-flask==1.2.0
coverage==7.2.3
gunicorn==20.1.0
